+++ cypress/support/commands.js
@@ -69,12 +69,36 @@ Cypress.Commands.add('isInViewport', { prevSubject: 'element' }, (element) => {
  * @example
  * cy.get('some-selector').paste('text');
  */
-Cypress.Commands.add('paste', {prevSubject: true}, (element, pastePayload) => {
+Cypress.Commands.add('paste', { prevSubject: true }, (element, pastePayload) => {
   const pasteEvent = Object.assign(new Event('paste', { bubbles: true, cancelable: true }), {
       clipboardData: {
           getData: () => pastePayload,
       },
   });
   element[0].dispatchEvent(pasteEvent);
-  return element
+  cy.wrap(element[0]).type('{insert}' + pastePayload);
 });
+
+/**
+ * Command changes default behavior of cy.window().blur() event
+ * so that it simulates blur event on last active element
+ */
+Cypress.Commands.add('windowBlur', {prevSubject: true}, (element) => {
+  cy.window()
+    .blur().then(() => {
+      element.blur();
+      return element;
+    })
+})
+
+/**
+ * Command changes default behavior of cy.window().focus() event
+ * so that it simulates returning focus to last active element
+ */
+Cypress.Commands.add('windowFocus', {prevSubject: true}, (element) => {
+  cy.window()
+    .focus().then(() => {
+      element.focus();
+      return element;
+    })
+})