+++ demo/CypressLayout.tsx
@@ -11,14 +11,17 @@ import {
   DateRange,
   DropDownSelect,
   DropZone,
+  FileDrop,
   Input,
   MaskedInput,
   Modal,
   VStepper,
   Wizard,
   DropDown,
+  DropDownLink,
   RadioGroup,
-  MultiSelect
+  MultiSelect,
+  ToolTip
 } from './cypress';
 import { Styles } from './components/Styles';
 
@@ -37,6 +40,9 @@ export const CypressLayout = () => (
         </Route>
         <Route path="/cypress/checkbox">
           <CheckBox />
+        </Route>
+        <Route path="/cypress/filedrop">
+          <FileDrop />
         </Route>
         <Route path="/cypress/input">
           <Input />
@@ -68,11 +74,17 @@ export const CypressLayout = () => (
         <Route path="/cypress/dropdown">
           <DropDown />
         </Route>
+        <Route path="/cypress/dropdownlink">
+          <DropDownLink />
+        </Route>
         <Route path="/cypress/radio-group">
           <RadioGroup />
         </Route>
         <Route path="/cypress/multi-select">
           <MultiSelect />
+        </Route>
+        <Route path="/cypress/tooltip">
+          <ToolTip />
         </Route>
       </Switch>
     </L.Div>