+++ demo/components/Tooltip.tsx
@@ -17,7 +17,7 @@ export const Tooltip = () => {
   ];
 
   return (
-    <L.Div _demoStory>
+    <L.Div style={{ position: 'relative', paddingBottom: '50px' }} _demoStory>
       <L.H4 _title>Tooltip</L.H4>
       <br />
       <L.ButtonGroup
@@ -46,19 +46,75 @@ export const Tooltip = () => {
       <br />
       {display && (
         <L.Div _tooltipDemo>
-          <L.Tooltip isOpen={open} position="top" title="Tooltip at top top top top top top top top top top top top top top top top">
+          <L.Tooltip
+            isOpen={open}
+            position="top"
+            title="Tooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top topTooltip at top top top top top top top top top top top top top top top top"
+          >
             <L.Button _tipTop>Top</L.Button>
           </L.Tooltip>
           <L.Div _tooltipLeftRight>
-            <L.Tooltip isOpen={open} position="left" title="Tooltip at left left left left left left left left left left left left">
+            <L.Tooltip
+              isOpen={open}
+              position="left"
+              title="Tooltip at left left left left left left left left left left left left"
+            >
               <L.Button _tipLeft>Left</L.Button>
             </L.Tooltip>
-            <L.Tooltip isOpen={open} position="right" title="Tooltip at right right right right right right right right right right">
+            <L.Tooltip
+              isOpen={open}
+              position="right"
+              title="Tooltip at right right right right right right right right right right"
+            >
               <L.Button _tipRight>Right</L.Button>
             </L.Tooltip>
           </L.Div>
-          <L.Tooltip isOpen={open} position="bottom" title="Tooltip at bottom bottom bottom bottom bottom bottom bottom bottom">
-            <L.Button onClick={() => setDisplay(!display)} _tipBottom>Bottom</L.Button>
+          <L.Tooltip
+            isOpen={open}
+            position="bottom"
+            title="Tooltip at bottom bottom bottom bottom bottom bottom bottom bottom"
+          >
+            <L.Button _tipBottom>
+              Bottom
+            </L.Button>
+          </L.Tooltip>
+          <L.Tooltip
+            isOpen={open}
+            title="Tooltip without positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning"
+          >
+            <L.Button style={{ position: "absolute", left: 0, bottom: "50px" }}>
+              defines the right side
+            </L.Button>
+          </L.Tooltip>
+          <L.Tooltip
+            isOpen={open}
+            title="Tooltip without positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning"
+          >
+            <L.Button
+              style={{ position: 'absolute', right: 0, bottom: '50px' }}
+            >
+              defines the left side
+            </L.Button>
+          </L.Tooltip>
+          <L.Tooltip
+            isOpen={open}
+            title="Tooltip without positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning"
+          >
+            <L.Button
+              style={{ position: 'absolute', left: '140px', bottom: '10px' }}
+            >
+              no positioning
+            </L.Button>
+          </L.Tooltip>
+          <L.Tooltip
+            isOpen={open}
+            title="Tooltip without positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning positioning"
+          >
+            <L.Button
+              style={{ position: 'absolute', right: '140px', bottom: '10px' }}
+            >
+              no positioning
+            </L.Button>
           </L.Tooltip>
         </L.Div>
       )}